+++ src/frontend/apps/e2e/__tests__/app-desk/config.spec.ts
@@ -23,6 +23,7 @@ test.describe('Config', () => {
         ['fr-fr', 'French'],
       ],
       FEATURES: { TEAMS: true },
+      RELEASE: 'NA',
     });
   });